fix timestamp sign problem

+++ bundler/src/main/java/io/ib67/sfcraft/bundler/EntryInputStream.java
@@ -25,8 +25,12 @@ public class EntryInputStream extends FilterInputStream {
     public BundleEntry readEntry() throws IOException {
         var magic = (short) (in.read() << 8 | in.read()); // u16
         if (magic != BundleEntry.MAGIC) throw new IOException("Invalid magic number: "+Long.toHexString(magic));
+        // every byte must be widened to long before shifting: an int shift such as
+        // `in.read() << 24` goes negative when the byte is >= 0x80 and sign-extends into the
+        // high 32 bits, corrupting the timestamp (which then makes the entry look "too old"
+        // and get silently skipped on restore).
         var time = (long) in.read() << 56 | (long) in.read() << 48 | (long) in.read() << 40 | (long) in.read() << 32 |
-                (long) in.read() << 24 | in.read() << 16 | in.read() << 8 | in.read();
+                (long) in.read() << 24 | (long) in.read() << 16 | (long) in.read() << 8 | (long) in.read();
         var attribute = VarInts.readVarInt(in);
         var segmentLen = VarInts.readVarLong(in);
         var lenName = VarInts.readVarInt(in);
